fix: PostgreSQL-compatible SQL types in all Flyway migrations; MODE=PostgreSQL in H2 test URLs
- DOUBLE → DOUBLE PRECISION in V13, V1000, V1001, V1002
- Added MODE=PostgreSQL to all H2 test JDBC URLs — H2 now rejects non-standard
  types at test time, turning silent failures into immediate build failures
- Removed Flyway bypass from PostgresTestResource — migrations now run on real
  PostgreSQL without workarounds
- All 678 tests passing (605 runtime + 73 reports)

Closes #146, Refs #104

+++ runtime/src/main/resources/db/migration/V13__confidence_score.sql
@@ -3,4 +3,4 @@
 -- Nullable Double (0.0–1.0). Null means the WorkItem was not created by an AI
 -- agent or no confidence metadata was provided.
 
-ALTER TABLE work_item ADD COLUMN confidence_score DOUBLE;
+ALTER TABLE work_item ADD COLUMN confidence_score DOUBLE PRECISION;
